+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
 ## [1.0.6] - 2022 Dec 1
-- Added {path} to get the full path of the current file
-- Added {selection} to get the selected text
+- Added {path} to command string to get the full path of the current file
+- Added {selection} to command string to get the selected text
+- Added "terminal" to individual/default commands to choose terminal name
 - Fixed issue with command being cut off with WSL, stop/logout, and clear
 
 ## [1.0.5] - 2022 Feb 8